Get rid of AutoCompletionListener infrastructure
This name AutoCompletionListener is already taken by
org.fife.ui.autocomplete, and it means something else.
Let's not conflate the terminology here.

For now, let's keep things simple, just registering and using the
JythonAutoCompletions directly, rather than trying to make it extensible
via a non-SciJava-based mechanism. If we need extensibility of
autocompletions later, we can introduce a new SciJava plugin type in the
script-editor component, and then rework this component to use it.